Make the wizard a component the next flow can use
The catalog wizard carried its own dialog, header, progress rail and
footer. That chrome is the same for any multi-step flow, so it moves to the
design system as Wizard, with the rail beside it as WizardProgress.

Deliberately a shell: which step is open and whether the flow may advance
stay with the host, because per-step validation differs enough between
flows that a shared guard would fit the first caller and fight the second.

Named Progress rather than Status — status means record lifecycle
everywhere else here, and this shows position in a sequence.
